+++ Core/Matcher/ContentTypeMatcher.php
@@ -19,6 +19,7 @@ class ContentTypeMatcher extends RepositoryMatcher implements KeyMatcherInterfac
     //const MATCH_CONTENTTYPE_REMOTE_ID = 'contenttype_remote_id';
 
     protected $allowedConditions = array(
+        self::MATCH_ALL,
         self::MATCH_CONTENTTYPE_ID, self::MATCH_CONTENTTYPE_IDENTIFIER, //self::MATCH_CONTENTTYPE_REMOTE_ID,
         // aliases
         'id', 'identifier', // 'remote_id'
@@ -60,6 +61,9 @@ class ContentTypeMatcher extends RepositoryMatcher implements KeyMatcherInterfac
                 /*case 'remote_id':
                 case self::MATCH_CONTENTTYPE_REMOTE_ID:
                     return new ContentTypeCollection($this->findContentTypesByRemoteId($values));*/
+
+                case self::MATCH_ALL:
+                    return new ContentTypeCollection($this->findAllContentTypes());
             }
         }
     }
@@ -118,6 +122,23 @@ class ContentTypeMatcher extends RepositoryMatcher implements KeyMatcherInterfac
             // return unique contents
             $contentType = $this->repository->getContentTypeService()->loadContentTypeByRemoteId($contentTypeRemoteId);
             $contentTypes[$contentType->id] = $contentType;
+        }
+
+        return $contentTypes;
+    }
+
+    /**
+     * @return ContentType[]
+     */
+    protected function findAllContentTypes()
+    {
+        $contentTypes = [];
+
+        $contentTypeService = $this->repository->getContentTypeService();
+        foreach ($contentTypeService->loadContentTypeGroups() as $contentTypeGroup) {
+            foreach ($contentTypeService->loadContentTypes($contentTypeGroup) as $contentType) {
+                $contentTypes[$contentType->id] = $contentType;
+            }
         }
 
         return $contentTypes;